FinalLast

+++ Final_Task/app/Http/Controllers/Client/CheckOutController.php
@@ -4,6 +4,10 @@ namespace App\Http\Controllers\Client;
 
 use App\Http\Controllers\Controller;
 use App\Models\Disposal;
+use App\Models\Order;
+use App\Models\OrderItem;
+use App\Models\ProductsImg;
+use App\Models\User;
 use Gloudemans\Shoppingcart\Facades\Cart;
 use Illuminate\Http\Request;
 
@@ -19,7 +23,7 @@ class CheckOutController extends Controller
         $request->validate([
             'fname' => 'required',
             'lname' => 'required',
-            'cname' => 'nullable',
+            'country' => 'required',
             'email' => 'required|email',
             'phone' => 'required',
             'address' => 'required',
@@ -28,17 +32,56 @@ class CheckOutController extends Controller
             'code' => 'nullable',
         ]);
 
-
+        $user = User::find(auth()->user()->id);
         $carts = Cart::content();
-        $disposal = new Disposal();
-        $disposal->user_id = auth()->user()->id;
-        $disposal->total = Cart::subtotal();
-        $disposal->save();
+        // $disposal = new Disposal();
+        // $disposal->user_id = $user->id;
+        // $disposal->total = Cart::subtotal();
+        // $disposal->save();
 
-        Cart::destroy();
+        $order = Order::create([
+            'total_count'=> Cart::count(),
+            'total_price'=> Cart::subtotal(),
+            'user_id'=> $user->id,
+            'country'=> $request->country,
+            'city'=> $request->city,
+            'address'=> $request->address,
+            'state'=> $request->state,
+            'zip_code'=> $request->code,
+            'order_number'=> uniqid(),
+        ]);
+
+        if ($order) {
+            foreach (Cart::content() as $key => $cart) {
+                $book = ProductsImg::find($cart->id);
+                if ($book) {
+                    if ($book->qty - $cart->qty >= 0) {
+                        $dataItem = [
+                            'book_id' => $book->id,
+                            'order_id' => $order->id,
+                            'qty' => $cart->qty,
+                            'price' => $book->price,
+                            'total_price' => $cart->price * $cart->qty,
+                        ];
+
+                        $orderItem = OrderItem::create($dataItem);
+                        if ($orderItem) {
+                            $book->qty -= $cart->qty;
+                            $book->save();
+                        }
+                    } else {
+                        return redirect()->back()->with('msgType', 'error')->with('message', 'Failed to create order! Because book id: ' . $book->id . ' count is not enough');
+                    }
+                } else {
+                    return redirect()->back()->with('msgType', 'error')->with('message', 'Failed to create order! Because book id: ' . $book->id . ' not found');
+                }
+            }
+
+            Cart::destroy();
+        }
 
 
-        return view('successfull', compact('disposal','carts'))->with('success', 'Sifariş uğurla yerinə yetirildi!');
+        return view('successfull', compact('order','carts'))->with('success', 'Sifariş uğurla yerinə yetirildi!');
     }
